requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.31.0
azure-storage-blob==12.23.1
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
msrest==0.7.1
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.1
packaging==24.1
pycparser==2.22
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.35
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
